requirements.txt
llama-index
llama-index-llms-openai
llama-index-embeddings-openai
python-dotenv
openai

